fix: SVG percentage in lg, using center coordinate of the conatiner
The SVG must use project-container-down-lg as the container if it exists as the closest wrapper, otherwise the animation does not work when < lg. However, when >= lg, project-container-down-lg must not have margin-bottom since the show-lg more buttons are in another row. There is no way to have margin-top while margin-bottom = 0 and the whole thing align vertically. By switching to using the center coordinate, it more or less works now.