Improve SVG compatability
Switch to <use xlink:href> and remove <svg style="rotate: 180deg">
for compatibility with Safari 11 and 12. Safari
10 is not supported due to the scope error and the lack of support
for the ... operator.

The scope error (SyntaxError: Cannot declare a let variable twice:
'e') can be fixed by supplying --safari10 to terser when doing the
mangling. But the ... operator is built in Bootstrap and has no easy
workaround.